+++ buildpacks/dotnet/tests/dotnet_publish_test.rs
@@ -106,6 +106,21 @@ fn test_dotnet_publish_with_project_toml_configuration() {
 
 #[test]
 #[ignore = "integration test"]
+fn test_solution_detection_with_multiple_workspace_root_solutions() {
+    TestRunner::default().build(
+        default_build_config("tests/fixtures/multiple_solutions"),
+        |context| {
+            assert_empty!(context.pack_stderr);
+            assert_contains!(
+                context.pack_stdout,
+                "- Detected .NET solution: `/workspace/bar.sln`"
+            );
+        },
+    );
+}
+
+#[test]
+#[ignore = "integration test"]
 fn test_dotnet_publish_process_registration_with_procfile() {
     TestRunner::default().build(
         default_build_config("tests/fixtures/basic_web_9.0_with_procfile"),