+++ buildpacks/dotnet/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Solution file paths in the root directory are now consistently sorted by the buildpack. ([#311](https://github.com/heroku/buildpacks-dotnet/pull/311))
 
+### Added
+
+- Support for configuring the solution to build and publish in `project.toml`. ([#310](https://github.com/heroku/buildpacks-dotnet/pull/310))
+
 ## [0.7.0] - 2025-08-27
 
 ### Added